Refactor CLI module into subpackage
