+++ fractal_server/cli/_start.py
@@ -1,0 +1,15 @@
+import uvicorn
+
+
+def start(
+    *,
+    host: str,
+    port: int,
+    reload: bool,
+):
+    uvicorn.run(
+        "fractal_server.main:app",
+        host=host,
+        port=port,
+        reload=reload,
+    )
